requirements.txt
apache-airflow==3.0.4
minio==7.2.16
requests==2.32.4
dotenv==0.9.9